+++ lib/flustars.dart
@@ -1,9 +1,10 @@
 library flustars;
 
-export 'src/ui/my_app_bar.dart';
-
 export 'package:common_utils/common_utils.dart';
 
+export 'src/ui/my_app_bar.dart';
+
+export 'src/directory_util.dart';
 export 'src/screen_util.dart';
 export 'src/widget_util.dart';
 export 'src/sp_util.dart';